Add NCSA delta proposal
Note that the PDF is "out of sync" with the markdown source. This is
because I didn't save the exact source code for the submitted PDF, and then
I continued to edit the source code. Therefore, the PDF reflects what is
actually submitted, while the source MD represents my current thinking.